Update dependencies; increase size of _seqno; restore Java 8 compatibility

+++ src/main/java/com/arakelian/elastic/model/DocumentId.java
@@ -37,7 +37,7 @@ public interface DocumentId {
     @Nullable
     @JsonProperty("_seq_no")
     @JsonView(Version6.class)
-    public Integer getSeqNo();
+    public Long getSeqNo();
 
     @Nullable
     @JsonProperty("_primary_term")
